// src/interfaces/models/ngayphep.model.ts
export interface INgayPhepModel{
    maNhanVien: number;
    nam: number;
    thang1: number;
    thang2: number;
    thang3: number;
    thang4: number;
    thang5: number;
    thang6: number;
    thang7: number;
    thang8: number;
    thang9: number;
    thang10: number;
    thang11: number;
    thang12: number;
    ngayPhepTichLuy: number;
    ngayPhepDaSuDung: number;
    ngayPhepConLai: number;
    ngayCapNhat: Date;
}